Created CWM::Page, CWM::Pager
by copying CWM::Tab, CWM::Tabs

A {Pager} contains several {Page}s and makes only one visible at a time.
A page widget is a group of widgets that are contained within it.

TreePager and Tabs will be subclasses of Pager.
