+++ library/cwm/src/Makefile.am
@@ -14,6 +14,8 @@ ylibdir = @ylibdir@/cwm
 ylib_DATA = \
   lib/cwm/abstract_widget.rb \
   lib/cwm/custom_widget.rb \
+  lib/cwm/page.rb \
+  lib/cwm/pager.rb \
   lib/cwm/tabs.rb \
   lib/cwm/tree.rb \
   lib/cwm/widget.rb
